set polygon creator to default nhood maker; removed all dependancies upon experimental map app - integrated new functionality into other apps.
