+++ map/templates/map/polygon.html
@@ -16,7 +16,7 @@
 
 {% endblock head %}
 {% block content %}
-    <form action="{% url 'index' %}" method="get">
+    <form action="#" method="get">
     <!-- {% csrf_token %} -->
         <div class="wrapper">
             <!-- Sidebar -->
