+++ neighborhood/templates/neighborhood/json_creator.html
@@ -1,0 +1,17 @@
+{% extends "base.html" %}
+{% load crispy_forms_tags %}
+{% block content %}
+<div class="content-section">
+  <form method="post">
+    {% csrf_token %}
+    <fieldset class="form-group">
+      <legend class="border-bottom mb-4">Define New Neighborhood - Text Option</legend>
+      {{ form|crispy }}
+    </fieldset>
+    <div class="form-group">
+      <button class="btn btn-outline-info" type="submit">Create</button>
+      <a href="{% url 'neighborhood' %}" class="btn btn-outline-dark">Return to Map Creator</a>
+    </div>
+  </form>
+</div>
+{% endblock content %}
