+++ neighborhood/templates/neighborhood/nhood_index.html
@@ -46,6 +46,7 @@
   <br>
   <div class="form-group">
     <button type="submit" name="nhood-create-button" class="btn btn-outline-info">Create</button>
+    <a href="{% url 'json_creator' %}" class="btn btn-outline-dark">Alternate - Create With JSON</a>
   </div>
 </form>
 {% endblock form %}
