+++ sample/templates/sample/sampling_index.html
@@ -1,0 +1,8 @@
+{% extends "index.html" %}
+{% load static %}
+
+{% block heading %}
+<h1 class="jumbotron-heading">Neighborhood Sampling Index</h1>
+<br>
+<p class=help-text>Select from one of the following regions:</p>
+{% endblock heading %}
